+++ daemon/daemon.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/zeroibot/pack/clock"
 	"github.com/zeroibot/pack/dict"
+	"github.com/zeroibot/pack/ds"
 	"github.com/zeroibot/pack/io"
 )
 
@@ -40,24 +41,24 @@ type Config struct {
 
 // LoadConfig loads a Daemon Config which is expected to follow the required structure.
 // This checks if any Interval values are 0 (invalid).
-func LoadConfig[T any](path string) (*T, error) {
+func LoadConfig[T any](path string) ds.Result[T] {
 	result := io.ReadJSON[T](path)
 	if result.IsError() {
-		return nil, result.Error()
+		return result
 	}
 	cfg := new(result.Value())
 	cfgMap, err := dict.FromStruct[map[string]int](cfg)
 	if err != nil {
-		return nil, err
+		return ds.Error[T](err)
 	}
 	for key := range cfgMap {
 		for cfgKey, value := range cfgMap[key] {
 			if value == 0 {
-				return nil, fmt.Errorf("invalid daemon %s.%s: %d", key, cfgKey, value)
+				return ds.Error[T](fmt.Errorf("invalid daemon %s.%s: %d", key, cfgKey, value))
 			}
 		}
 	}
-	return cfg, nil
+	return result
 }
 
 // Run runs a task every given interval, where TimeScale = time.Hour, time.Minute, time.Second