+++ model/model.go
@@ -2,6 +2,7 @@
 package model
 
 import (
+	"github.com/zeroibot/pack/ds"
 	"github.com/zeroibot/pack/dyn"
 	"github.com/zeroibot/pack/qb"
 )
@@ -15,10 +16,10 @@ type Schema[T any] struct {
 }
 
 // NewSchema creates a new Schema and registers its type to qb
-func NewSchema[T any](this *qb.Instance, structRef *T, table string) (*Schema[T], error) {
+func NewSchema[T any](this *qb.Instance, structRef *T, table string) ds.Result[*Schema[T]] {
 	err := qb.AddType(this, structRef)
 	if err != nil {
-		return nil, err
+		return ds.Error[*Schema[T]](err)
 	}
 	schema := new(Schema[T]{
 		Name:     dyn.TypeName(structRef),
@@ -27,28 +28,28 @@ func NewSchema[T any](this *qb.Instance, structRef *T, table string) (*Schema[T]
 		Reader:   qb.FullRowReader(this, structRef),
 		instance: this,
 	})
-	return schema, nil
+	return ds.NewResult(schema, nil)
 }
 
 // NewSharedSchema creates a new shared Schema (no table)
-func NewSharedSchema[T any](this *qb.Instance, structRef *T) (*Schema[T], error) {
+func NewSharedSchema[T any](this *qb.Instance, structRef *T) ds.Result[*Schema[T]] {
 	return NewSchema(this, structRef, "")
 }
 
 // AddSchema adds a new Schema and adds error to the error list if any
 func AddSchema[T any](this *qb.Instance, structRef *T, table string, errs []error) *Schema[T] {
-	schema, err := NewSchema(this, structRef, table)
-	if err != nil {
-		errs = append(errs, err)
+	result := NewSchema(this, structRef, table)
+	if result.IsError() {
+		errs = append(errs, result.Error())
 	}
-	return schema
+	return result.Value()
 }
 
 // AddSharedSchema adds a new shared Schema (no table) and adds error to the error list if any
 func AddSharedSchema[T any](this *qb.Instance, structRef *T, errs []error) *Schema[T] {
-	schema, err := NewSharedSchema(this, structRef)
-	if err != nil {
-		errs = append(errs, err)
+	result := NewSharedSchema(this, structRef)
+	if result.IsError() {
+		errs = append(errs, result.Error())
 	}
-	return schema
+	return result.Value()
 }